Expose dtaTree in Tree component and clean some things up
